2주차 백엔드 Merge Request 요청합니다

See merge request ai_track/class_06/web_project/team13/13team_web_project_backend!4

+++ src/services/product.service.js
@@ -1,6 +1,7 @@
 import {productRepository} from "../db";
 import {CustomError} from "../middlewares/filter";
 import {categoryService} from "./index";
+import commonErrors from "../middlewares/filter/error/commonErrors";
 
 
 class ProductService {
@@ -11,14 +12,11 @@ class ProductService {
 
     // 상품 등록
     async addProduct(productInfo) {
-        // 입력된 카테고리를 카테고리 DB에서 검색 
+        // 입력된 카테고리를 카테고리 DB 에서 검색
         const category = await categoryService.getCategoryByName(
             productInfo.category
         );
 
-        if (!category) {
-            throw new CustomError(404, "카테고리를 찾을수 없습니다.");
-        }
 
         productInfo.category = category;
 
@@ -56,10 +54,9 @@ class ProductService {
     async getProductById(productId) {
         //  해당 상품이 db에 존재하는지 확인
         const product = await this.productRepository.findById(productId);
+
         if (!product) {
-            throw new CustomError(
-                404, "해당 상품이 존재하지 않습니다. 다시 확인하고 시도해 주세요."
-            );
+            throw new CustomError(404, commonErrors.resourceNotFoundError)
         }
         return product;
     }
@@ -67,7 +64,11 @@ class ProductService {
     // 상품의 orderCount 증감
     async increaseOrderCount(productId, amount) {
 
-        await this.getProductById(productId);
+        const product = await this.productRepository.findById(productId);
+
+        if (!product) {
+            throw new CustomError(404, commonErrors.resourceNotFoundError);
+        }
 
         const orderproduct = await this.productRepository.update(productId, {
             $inc: {orderCount: amount},
@@ -77,17 +78,17 @@ class ProductService {
     }
 
     // 상품 정보 수정
-    async setProduct(productId, update) {
+    async setProduct(productId, updateInfo) {
 
-        let NewCategory = update
+        let updatedProduct = updateInfo
 
         // 입력된 카테고리를 카테고리 DB에서 검색 
-        NewCategory.category = await categoryService.getCategoryByName(
-            update.category
+        updatedProduct.category = await categoryService.getCategoryByName(
+            updateInfo.category
         );
 
 
-        const product = await this.productRepository.update(productId, NewCategory);
+        const product = await this.productRepository.update(productId, updatedProduct);
 
         return product;
     }
@@ -95,7 +96,11 @@ class ProductService {
     // 상품 정보 삭제
     async deleteProduct(productId) {
 
-        await this.getProductById(productId);
+        const product = await this.getProductById(productId);
+
+        if (!product) {
+            throw new CustomError(404, commonErrors.resourceNotFoundError);
+        }
 
         const deletedproduct = await this.productRepository.delete(productId);
